pyproject.toml
[project]
name = "mcp-ssh-gateway"
version = "0.1.0"
description = "MCP-compatible SSH Gateway agent"
authors = [{name = "Oliver Lanz"}]
license = "Apache-2.0"
dependencies = [
  "mcp-agent",
  "paramiko",
]

[tool.setuptools]
packages = ["agent"]
